Add ‘<&&>’ and ‘<||>’ combinators
These are often useful for writing lexing predicates (`Char -> Bool`),
as well as for writing verifiers (`Parser Bool`). I always find myself
rewriting them, and this package seems like a good home for them.

This adds both applicative and monadic forms, and a bit of documentation
to clarify their different short-circuiting behavior.

Since the possibility space for `Bool` is so small, the included tests
are exhaustive unit tests, instead of using random sampling.